+++ frontend/app/components/modules/HorizontalCard/HorizontalCard.tsx
@@ -10,7 +10,7 @@ type Props = {
 export const HorizontalCard = ({ children, className, title }: Props) => {
   return (
     <div className={styles.horizontalCard}>
-      <h2 className={styles.title}>{title}</h2>
+      {title && <h2 className={styles.title}>{title}</h2>}
       <div className={`${styles.card} ${className}`}>{children}</div>
     </div>
   )